Split mount class to its own file.
