+++ engine/src/cmd/mount.cpp
@@ -1,3 +1,28 @@
+/**
+ * mount.cpp
+ *
+ * Copyright (C) Daniel Horn
+ * Copyright (C) 2020 pyramid3d, Stephen G. Tuggy, and other Vega Strike
+ * contributors
+ *
+ * https://github.com/vegastrike/Vega-Strike-Engine-Source
+ *
+ * This file is part of Vega Strike.
+ *
+ * Vega Strike is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Vega Strike is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Vega Strike.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 #include "unit_generic.h"
 #include "missile.h"
 #include "beam.h"
@@ -476,3 +501,30 @@ void Mount::ReplaceSound()
 {
     sound = AUDCreateSound( sound, false );     //copy constructor basically
 }
+
+void Mount::Activate( bool Missile )
+{
+    if ( type->isMissile() == Missile )
+        if (status == INACTIVE)
+            status = ACTIVE;
+}
+
+///Sets this gun to inactive, unless unchosen or destroyed
+void Mount::DeActive( bool Missile )
+{
+    if ( type->isMissile() == Missile )
+        if (status == ACTIVE)
+            status = INACTIVE;
+}
+
+void Mount::SetMountPosition( const Vector &v )
+{
+    pos = v;
+}
+
+void Mount::SetMountOrientation( const Quaternion &t )
+{
+    orient = t;
+}
+
+
